move archive tests to archive folder
