+++ x-pack/plugins/fleet/server/services/epm/registry/index.test.ts
@@ -5,55 +5,9 @@
  * 2.0.
  */
 
-import type { AssetParts } from '../../../types';
-import { getBufferExtractor, getPathParts, untarBuffer, unzipBuffer } from '../archive';
-
 import { splitPkgKey } from '.';
 
-describe('getPathParts', () => {
-  const testPaths = [
-    {
-      path: 'foo-1.1.0/service/type/file.yml',
-      assetParts: {
-        dataset: undefined,
-        file: 'file.yml',
-        path: 'foo-1.1.0/service/type/file.yml',
-        pkgkey: 'foo-1.1.0',
-        service: 'service',
-        type: 'type',
-      },
-    },
-    {
-      path: 'iptables-1.0.4/kibana/visualization/683402b0-1f29-11e9-8ec4-cf5d91a864b3-ecs.json',
-      assetParts: {
-        dataset: undefined,
-        file: '683402b0-1f29-11e9-8ec4-cf5d91a864b3-ecs.json',
-        path: 'iptables-1.0.4/kibana/visualization/683402b0-1f29-11e9-8ec4-cf5d91a864b3-ecs.json',
-        pkgkey: 'iptables-1.0.4',
-        service: 'kibana',
-        type: 'visualization',
-      },
-    },
-    {
-      path: 'coredns-1.0.1/data_stream/stats/fields/coredns.stats.yml',
-      assetParts: {
-        dataset: 'stats',
-        file: 'coredns.stats.yml',
-        path: 'coredns-1.0.1/data_stream/stats/fields/coredns.stats.yml',
-        pkgkey: 'coredns-1.0.1',
-        service: '',
-        type: 'fields',
-      },
-    },
-  ];
-  test('testPathParts', () => {
-    for (const value of testPaths) {
-      expect(getPathParts(value.path)).toStrictEqual(value.assetParts as AssetParts);
-    }
-  });
-});
-
-describe('splitPkgKey tests', () => {
+describe('splitPkgKey', () => {
   it('throws an error if there is nothing before the delimiter', () => {
     expect(() => {
       splitPkgKey('-0.0.1-dev1');
@@ -84,41 +38,3 @@ describe('splitPkgKey tests', () => {
     expect(pkgVersion).toBe('0.13.0-alpha.1+abcd');
   });
 });
-
-describe('getBufferExtractor called with { archivePath }', () => {
-  it('returns unzipBuffer if `archivePath` ends in .zip', () => {
-    const extractor = getBufferExtractor({ archivePath: '.zip' });
-    expect(extractor).toBe(unzipBuffer);
-  });
-
-  it('returns untarBuffer if `archivePath` ends in .gz', () => {
-    const extractor = getBufferExtractor({ archivePath: '.gz' });
-    expect(extractor).toBe(untarBuffer);
-    const extractor2 = getBufferExtractor({ archivePath: '.tar.gz' });
-    expect(extractor2).toBe(untarBuffer);
-  });
-
-  it('returns `undefined` if `archivePath` ends in anything else', () => {
-    const extractor = getBufferExtractor({ archivePath: '.xyz' });
-    expect(extractor).toEqual(undefined);
-  });
-});
-
-describe('getBufferExtractor called with { contentType }', () => {
-  it('returns unzipBuffer if `contentType` is `application/zip`', () => {
-    const extractor = getBufferExtractor({ contentType: 'application/zip' });
-    expect(extractor).toBe(unzipBuffer);
-  });
-
-  it('returns untarBuffer if `contentType` is `application/gzip`', () => {
-    const extractor = getBufferExtractor({ contentType: 'application/gzip' });
-    expect(extractor).toBe(untarBuffer);
-  });
-
-  it('returns `undefined` if `contentType` ends in anything else', () => {
-    const extractor = getBufferExtractor({ contentType: '.xyz' });
-    expect(extractor).toEqual(undefined);
-  });
-});
-
-describe('fetchFindLatestPackageOrUndefined', () => {});
